requirements.txt
pandas>=2.1
numpy>=1.26
pyyaml>=6.0
beautifulsoup4>=4.12
lxml>=5.0
tqdm>=4.66
matplotlib>=3.8
scikit-learn>=1.4
